+++ examples/opencode/plugins/xmpp-bridge.js
@@ -57,7 +57,7 @@
  */
 
 export const XmppBridgePlugin = async ({ client, directory, $ }) => {
-   const PLUGIN_VERSION = "0.8.25"
+   const PLUGIN_VERSION = "0.8.26"
   const pluginRef = (() => {
     try {
       // eslint-disable-next-line no-undef